+++ lib/scrambled_eggs.rb
@@ -8,15 +8,19 @@ class ScrambledEggs
   # _algorithm_ ::  Algorithm (ex. 'aes-256-cbc')
   # _salt_      ::  Salt (8 bytes)
   # _key_       ::  Crypt key
+  # _key_file_  ::  Crypt key file (exclude key)
   #
   # return      ::  ScrambledEggs object
-  def initialize( algorithm: 'aes-256-cbc', salt: nil, key: nil )
+  def initialize( algorithm: 'aes-256-cbc', salt: nil, key: nil, key_file: nil )
     @@algorithm = algorithm
     @@salt = salt != nil ? salt : OpenSSL::Random.random_bytes( 8 )
     if key
       @@key = key
     else
-      @@key = File.open( '/etc/hostname' ).read
+      unless key_file
+        key_file = '/etc/hostname'
+      end
+      @@key = OpenSSL::Digest::SHA512.digest( Pathname.new( key_file ).binread )
     end
   end
 
@@ -44,6 +48,22 @@ class ScrambledEggs
     data = scrambled[ 8, scrambled.size ]
     cipher.pkcs5_keyivgen( @@key, salt )
     cipher.update( data ) + cipher.final
+  end
+
+  # Scramble (encrypt) file
+  #
+  # _path_  ::  File for scramble
+  def scramble_file( path )
+    pathname = Pathname.new( path )
+    pathname.binwrite( scramble( pathname.binread ) )
+  end
+
+  # Descramble (decrypt) file
+  #
+  # _path ::  File for descramble
+  def descramble_file( path )
+    pathname = Pathname.new( path )
+    pathname.binwrite( descramble( pathname.binread ) )
   end
 end