+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+### 0.1.0 / 1015-02-11  sanadan <jecy00@gmail.com>
+
+  * Add generate key from key file.
+  * Change default key to generate key from hostname file.
+  * Add scramble/descramble file function.
+
+
 ### 0.0.2 / 2015-02-07  sanadan <jecy00@gmail.com>
 
   * Fixed: Can't require in gemspec file
+
 
 ### 0.0.1 / 2015-02-06  sanadan <jecy00@gmail.com>